# docs/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Factory Ops API
  description: |
    工廠值班工作管理系統 REST API(v1.3)。
    - API-first,Web 與未來 Native App 共用同一組 API。
    - 認證採 JWT Bearer Token(access + refresh)。JWT claims 帶 `userId`、`accountName`、`rootOrgId`、`orgPath[]`、`roles[]`、`groupIds[]`、`orgManagerScopes[]`(衍生自 Organization.managerId 反查)。
    - **多租戶隔離**:所有資源(除 `Organization` 自身)歸屬於某個 `rootOrgId`,server 強制 scope。跨 root org 操作僅 `ADMIN` 系統角色可進行。
    - **時間欄位**:wire format 使用 ISO 8601 + offset(例 `2026-05-04T08:30:00+08:00`,Q-17 拍板);儲存層為 UTC `Instant`。
    - 列表端點預設使用 cursor-based pagination,支援欄位投影與增量同步。
    - **Organization 為 polymorphic 樹**(FAB → DIVISION → DEPARTMENT → SECTION),只有 leaf type 可以承載 Group / Project / Task(見 ADR-0004)。每節點有 **單一 `managerId`**(對應 ORG_MANAGER)+ **多位 `leaderIds[]`**(業務負責人列表)— 見 ADR-0010。
    - **Group 為平面結構**(屬唯一 leaf Organization),polymorphic by `type`(DEFAULT / LINE / TEAM / SHIFT…);Group 加 `settings.qa` 用於設定 QA 雙簽機制(見 ADR-0011)。
    - **Template 機制**:支援 GLOBAL(ADMIN 維護)與 ORG(各 Org 維護)兩種 scope;可 fork from GLOBAL,可直接實例化(見 ADR-0006)。
    - **跨層派工(v1.3 改寫)**:任一 Organization 節點的 manager 可直接派 ActionRequest 到其子孫中的 leaf Org(`targetOrgId` 必為 leaf);**v1.3 移除多層 relay 流程**(見 ADR-0008 v1.3)。
    - **Task QA 雙簽**:Task 建立時從所屬 Project.groupIds 對應 Group settings.qa snapshot policy;之後 Group settings 變更不影響該 Task;Review action 走 `POST /tasks/{taskId}/review`(見 ADR-0011)。
    - **HR 整合**:User 為 HR User 的本地投影,以 `accountName` 對 HR Mock REST 查詢(見 ADR-0007 v1.3 Amendment)。
...
  version: 1.3.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /orgs:
    get:
      tags: [Organizations]
      summary: 列出 Organization 節點(僅當前 user 所屬 root 樹的節點;ADMIN 跨樹)
      description: |
        支援多種過濾:
        - `type=FAB|DIVISION|DEPARTMENT|SECTION|...`
        - `parentId=<id>`(取直接子節點;傳 `null` 取 root)
        - `leafOnly=true`(只回 type ∈ leafTypes 的節點)
        - `underOrgId=<id>`(回該節點及其子孫;server 沿樹計算)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: type
          schema: { type: string }
          description: 過濾 Organization 型態
        - in: query
          name: parentId
          schema: { type: string, nullable: true }
          description: 過濾 parentId;傳 `null` 取 root
        - in: query
          name: leafOnly
          schema: { type: boolean, default: false }
          description: 只回 leaf type(可承載工作的節點)
        - in: query
          name: underOrgId
          schema: { type: string }
          description: 沿樹回傳該節點及其子孫
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: Organization 清單
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationPage' }
    post:
      tags: [Organizations]
      summary: 建立 Organization 節點
      description: |
        - 建立 root(無 parentId):僅 `ADMIN` 可。
        - 建立子節點(有 parentId):`ORG_ADMIN`(同 root)或 `ADMIN`。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganizationRequest' }
      responses:
        '201':
          description: 已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /orgs/{orgId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [Organizations]
      summary: 讀取單一 Organization 節點
      responses:
        '200':
          description: Organization
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Organizations]
      summary: 更新 Organization 節點(name / parentId / leaderId / settings(僅 root))
      description: 變更 parentId 會即時驗證 cycle 與深度上限,失敗回 409
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrganizationRequest' }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Organizations]
      summary: 軟刪除 Organization 節點
      description: |
        - 非 leaf 節點:必須先處理(移除/軟刪)所有子孫,否則 409
        - root:cascade soft-delete 整棵樹下所有資源
      responses:
        '204': { description: No Content }
        '409':
          $ref: '#/components/responses/Conflict'
  /orgs/{orgId}/transfer-manager:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    post:
      tags: [Organizations]
      summary: 設置 / 變更 / 清空 Organization 節點 manager(ORG_ADMIN / ADMIN;v1.3 取代 transfer-leader)
      description: |
        - 每個 Organization 節點最多一位 `managerId`(對應 ORG_MANAGER 角色)。
        - 傳 `newManagerId: null` 表示清空(該節點暫無 manager)。
        - 變更會寫入 `Organization.history[]` 並 emit `factory-ops.org.manager-transferred`。
        - 對應的 `User.orgManagerScopes[]` 衍生欄位由 reactor 同步維護。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newManagerId]
              properties:
                newManagerId:
                  type: string
                  nullable: true
                  description: 新 manager 的 userId;傳 null 表示清空
                reason: { type: string }
      responses:
        '200':
          description: 已變更
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '422':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /orgs/{orgId}/leaders:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [Organizations]
      summary: 列出該 Organization 節點的 leaders(0..N)
      description: |
        回傳 `leaderIds[]` 對應的 User snapshot;leader 為該節點的業務負責人,
        於 ActionRequest dispatch 時為 ownerId 預設候選(見 ADR-0010)。
      responses:
        '200':
          description: Leader 清單
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
    post:
      tags: [Organizations]
      summary: 新增一位 leader(ORG_ADMIN / ADMIN)
      description: |
        - 將 `userId` 加入該節點的 `leaderIds[]`;若已存在則 409。
        - user 必須與 Organization 同 `rootOrgId` 且 `active = true`,否則 422。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string }
                reason: { type: string }
      responses:
        '201':
          description: 已加入
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /orgs/{orgId}/leaders/{userId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
      - $ref: '#/components/parameters/UserIdParam'
    delete:
      tags: [Organizations]
      summary: 移除一位 leader(ORG_ADMIN / ADMIN)
      description: |
        - 將指定 `userId` 自 `leaderIds[]` 移除。
        - 若為最後一位 leader,允許移除(`leaderIds.length == 0`),但之後新派工到此 leaf 會 409 `target_org_no_leader`。
      responses:
        '204': { description: No Content }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgIdParam:
      in: path
      name: orgId
      required: true
      schema: { type: string }
      description: |
        Organization ID。在 `/orgs/{orgId}/...` 子路徑中,根據端點語意可能是 root org 或 任一節點:
        - `/orgs/{orgId}/groups`、`/orgs/{orgId}/project-templates` 等管理面 API:`orgId` 為 **rootOrgId**
        - `/orgs/{orgId}/dispatch-action-request`、`/orgs/{orgId}/transfer-leader`:`orgId` 為**特定節點**(可能任一層)
...
    UserIdParam:
      in: path
      name: userId
      required: true
      schema: { type: string }
    CursorParam:
      in: query
      name: cursor
      schema: { type: string }
      description: cursor-based pagination 游標
    LimitParam:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
...
    Forbidden:
      description: 無權限
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: 找不到資源
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 業務規則衝突(狀態流、唯一性)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: 輸入驗證失敗
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
  schemas:
...
    Problem:
      description: RFC 7807 problem details
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
      required: [title, status]
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  code: { type: string }
                  message: { type: string }
    PageInfo:
      type: object
      properties:
        nextCursor: { type: string, nullable: true }
        hasMore: { type: boolean }
...
    Role:
      type: string
      enum: [OPERATOR, SHIFT_LEAD, ENGINEER, QA, GROUP_ADMIN, GROUP_MANAGER, ORG_MANAGER, ORG_ADMIN, ADMIN]
      description: |
        - `OPERATOR` / `SHIFT_LEAD` / `ENGINEER` / `QA`:第一線執行
        - `GROUP_ADMIN`:Group 內 Template 與成員角色設定;Group settings(含 QA 雙簽)維護
        - `GROUP_MANAGER`:Group 經理,可指派 GROUP_ADMIN 代理;Group settings 最終設定權;權限 ⊇ GROUP_ADMIN
        - `ORG_MANAGER`:某 Organization 節點的唯一經理(衍生自 `Organization.managerId == userId`);可直接派 ActionRequest 到其子孫中的 leaf Org
        - `ORG_ADMIN`:該 root org 的最高權限;管理 Org 樹 / users / managers / leaders
        - `ADMIN`:系統管理員(SaaS 維運),唯一可維護 GLOBAL Templates
        > v1.3 注意:`ORG_MANAGER` 為衍生角色,不應由前端直接寫入 `User.roles[]`;由 server 從 `Organization.managerId` 反查並維護 `User.orgManagerScopes[]` 衍生欄位。
    OrganizationType:
      type: string
      description: Organization 多型 discriminator;內建型態,leaf 由 root settings.leafTypes 決定
      enum: [FAB, DIVISION, DEPARTMENT, SECTION]
...
    Organization:
      type: object
      properties:
        id: { type: string }
        rootOrgId: { type: string, description: 該節點所屬樹的 root;root 自己時等於 id }
        parentId: { type: string, nullable: true, description: null 僅 root }
        type:
          oneOf:
            - $ref: '#/components/schemas/OrganizationType'
            - type: string
              description: 自訂 type code(由 root settings 擴展)
        name: { type: string, description: 允許 Unicode 多語系混合輸入 }
        code: { type: string, description: 同 rootOrgId 內 unique }
        managerId:
          type: string
          nullable: true
          description: |
            該節點唯一經理(對應 ORG_MANAGER 角色;v1.3 新欄位)。
            為跨層 ActionRequest dispatch 授權的權威來源;`User.orgManagerScopes[]` 由此反查衍生。
        leaderIds:
          type: array
          description: |
            該節點業務負責人列表(0..N;v1.3 新欄位,取代 v1.2 leaderId 單值)。
            ActionRequest dispatch 預設 ownerId 從此集合中選(0 → 409,1 → 自動,N → 派工方指定)。
          items: { type: string }
        isLeaf: { type: boolean, description: 衍生欄位:type ∈ root.settings.leafTypes }
        depth: { type: integer, description: 衍生欄位:從 root 的距離(0 = root) }
        path: { type: string, description: 衍生欄位:`/<rootId>/.../<self>`(materialized,可選) }
        timezone: { type: string, example: Asia/Taipei, description: 僅 root 有效 }
        locale: { type: string, example: zh-TW, description: 僅 root 有效 }
        settings:
          allOf:
            - $ref: '#/components/schemas/OrgSettings'
          nullable: true
          description: 僅 root 有效
        createdAt: { type: string, format: date-time, description: ISO 8601 + offset }
        updatedAt: { type: string, format: date-time, description: ISO 8601 + offset }
        deletedAt: { type: string, format: date-time, nullable: true }
        schemaVersion: { type: integer }
      required: [id, rootOrgId, type, name, code]
    OrgSettings:
      type: object
      properties:
        orgMaxDepth:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          description: Organization 樹最大深度
        leafTypes:
          type: array
          items: { type: string }
          default: [SECTION]
          description: 哪些 type code 視為 leaf(可承載 Group / Project / Task)
        attachmentMaxBytes:
          type: integer
          format: int64
          example: 52428800
        extras:
          type: object
          additionalProperties: true
    CreateOrganizationRequest:
      type: object
      required: [type, name, code]
      properties:
        parentId:
          type: string
          nullable: true
          description: null 表示建立 root(僅 ADMIN);其他需 ORG_ADMIN(同 root)
        type: { type: string, description: FAB / DIVISION / DEPARTMENT / SECTION / 自訂 }
        name: { type: string, minLength: 1, maxLength: 120, description: 允許 Unicode 多語系混合 }
        code: { type: string, minLength: 1, maxLength: 64, pattern: '^[a-z0-9-]+$' }
        managerId:
          type: string
          nullable: true
          description: 建立時可選指定 manager(v1.3)
        leaderIds:
          type: array
          description: 建立時可選指定 leaders(0..N;v1.3)
          items: { type: string }
        timezone: { type: string, default: Asia/Taipei, description: 僅建 root 時有效 }
        locale: { type: string, default: zh-TW, description: 僅建 root 時有效 }
        settings:
          allOf:
            - $ref: '#/components/schemas/OrgSettings'
          nullable: true
          description: 僅建 root 時有效
        initialOrgAdmin:
          type: object
          description: 僅建 root 時提供;同步建立第一個 ORG_ADMIN 帳號
          required: [accountName]
          properties:
            accountName:
              type: string
              maxLength: 30
              description: 從 HR 服務拉取 profile
    UpdateOrganizationRequest:
      type: object
      description: |
        本端點不支援變更 `managerId` / `leaderIds[]`。
        - `managerId` 變更請使用 `POST /orgs/{orgId}/transfer-manager`
        - `leaderIds[]` 變更請使用 `POST /orgs/{orgId}/leaders` 與 `DELETE /orgs/{orgId}/leaders/{userId}`
      properties:
        name: { type: string }
        parentId: { type: string, nullable: true, description: 變更會做 cycle / 深度檢查 }
        type: { type: string }
        timezone: { type: string, description: 僅 root 有效 }
        locale: { type: string, description: 僅 root 有效 }
        settings:
          allOf:
            - $ref: '#/components/schemas/OrgSettings'
          description: 僅 root 有效
    OrganizationPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Organization' }
        pageInfo: { $ref: '#/components/schemas/PageInfo' }

    # ---------- Group(平面)----------
...
    User:
      type: object
      description: HR User 的本地投影;profile 欄位以 HR 為權威,roles / orgManagerScopes / GroupMembership 由本系統管理
      properties:
        id: { type: string }
        rootOrgId: { type: string }
        accountName:
          type: string
          maxLength: 30
          description: HR 識別字;`(rootOrgId, accountName)` unique
        employeeNo: { type: string, description: snapshot from HR }
        email: { type: string, format: email, description: snapshot from HR }
        displayName: { type: string, description: snapshot from HR }
        roles:
          type: array
          items: { $ref: '#/components/schemas/Role' }
        groupIds:
          type: array
          description: 衍生欄位:該 user 目前 active 屬於的 Group ID 集合
          items: { type: string }
        orgManagerScopes:
          type: array
          description: |
            **衍生 / cache 欄位(v1.3)**:該 user 是哪些 Organization 節點的 `managerId`。
            權威來源是 `Organization.managerId == userId` 反查;前端不可直接寫入(write-protected)。
            授權判定可即時從 Organization 反查作為防呆,本欄位為 RBAC 加速 cache。
          items: { type: string }
        primaryOrgPath:
          type: array
          description: 衍生欄位:該 user 所屬 leaf Org 從 root 的路徑
          items: { type: string }
        hrSyncedAt: { type: string, format: date-time, description: 最近一次 HR 同步時間 }
        active: { type: boolean, description: HR 標離職時 false }
        createdAt: { type: string, format: date-time }
      required: [id, rootOrgId, accountName, displayName, roles, active]
...